Update README.md
We can see that May was the best month to launch campaigns on the line graph, and January, June, July and October all had about the same number of failed campaigns launched.
The bar graph on the other hand, is just a representation of the outcomes of the Parent Categories. As we can see theater had the most success and music.